Separating Auth and Logger from Service
