+++ Sources/CompilerSwiftAI/Auth/Auth.swift
@@ -1,0 +1,5 @@
+//  Copyright © 2025 Compiler, Inc. All rights reserved.
+
+public protocol TokenManaging: Actor {
+    func getValidToken() async throws -> String
+}
