+++ Sources/CompilerSwiftAI/Auth/AuthError.swift
