+++ Sources/CompilerSwiftAI/Auth/AuthResponse.swift
